Update README.md
This project is a web application that utilizes the power of vanilla JavaScript to provide a seamless user experience. The main feature of the app is its ability to search for food recipes using the Themealdb API. The API returns a variety of data related to the searched recipe, such as ingredients, instructions, and nutritional information. The app then displays this information in an easy-to-read format, allowing users to quickly find the recipe they are looking for.

In addition to the search feature, the app also includes the ability for users to view details of the selected recipe. This includes a more in-depth look at the recipe's ingredients and instructions, as well as nutritional information. Users can also mark their favorite recipes for easy access in the future.

Overall, this project demonstrates your skills in working with APIs, JavaScript, and web development in general. The use of vanilla JavaScript allows for a lightweight and fast-performing application, without the need for any additional JavaScript frameworks. It's a great addition to your portfolio to show your ability to create a functional web application from scratch.